refactor(test): move JSON test to index.test.ts
