+++ src/index.test.ts
@@ -29,6 +29,7 @@ import {
   decodeKeyValue,
   encodeKey,
   encodeKeyValue,
+  hashData,
 } from './lib/utils';
 import { ERC725JSONSchema } from './types/ERC725JSONSchema';
 import {
@@ -47,6 +48,7 @@ import {
 import 'isomorphic-fetch';
 
 import { Schema } from '../test/generatedSchema';
+import { SUPPORTED_HASH_FUNCTION_STRINGS } from './lib/constants';
 
 const address = '0x0c03fba782b07bcf810deb3b7f0595024a444f4e';
 
@@ -58,7 +60,7 @@ describe('Running @erc725/erc725.js tests...', () => {
         // eslint-disable-next-line no-new
         new ERC725();
       },
-      (error) => error.message === 'Missing schema.',
+      (error: any) => error.message === 'Missing schema.',
     );
   });
 
@@ -69,7 +71,7 @@ describe('Running @erc725/erc725.js tests...', () => {
         // eslint-disable-next-line no-new
         new ERC725(mockSchema, address, { test: false });
       },
-      (error) =>
+      (error: any) =>
         error.message.indexOf('Incorrect or unsupported provider') >= -1,
     );
   });
@@ -468,5 +470,82 @@ describe('Running @erc725/erc725.js tests...', () => {
         });
       }
     }
+  });
+
+  it('should encode/decode JSON properly', () => {
+    const schema: ERC725JSONSchema[] = [
+      {
+        name: 'LSP3Profile',
+        key: '0x5ef83ad9559033e6e941db7d7c495acdce616347d28e90c7ce47cbfcfcad3bc5',
+        keyType: 'Singleton',
+        valueContent: 'JSONURL',
+        valueType: 'bytes',
+      },
+    ];
+
+    const myERC725 = new ERC725<Schema>(schema);
+
+    const json = {
+      name: 'rryter',
+      description: 'Web Developer located in Switzerland.',
+      profileImage: [
+        {
+          width: 1350,
+          height: 1800,
+          hashFunction: 'keccak256(bytes)',
+          hash: '0x229b60ea5b58e1ab8e6f1063300be110bb4fa663ba75d3814d60104ac6b74497',
+          url: 'ipfs://Qmbv9j6iCDDYJ1NXHTZnNHDJ6qaaKkZsf79jhUMFAXcfDR',
+        },
+        {
+          width: 768,
+          height: 1024,
+          hashFunction: 'keccak256(bytes)',
+          hash: '0x320db57770084f114988c8a94bcf219ca66c69421590466a45f382cd84995c2b',
+          url: 'ipfs://QmS4m2LmRpay7Jij4DCpvaW5zKZYy43ATZdRxUkUND6nG3',
+        },
+      ],
+      backgroundImage: [
+        {
+          width: 1024,
+          height: 768,
+          hashFunction: 'keccak256(bytes)',
+          hash: '0xbe2d39fe1e0b1911155afc74010db3483528a2b645dea8fcf47bdc34147769be',
+          url: 'ipfs://QmQ6ujfKSc91F44KtMe6WRTSCXoSdCjomQUy8hCUxHMr28',
+        },
+        {
+          width: 640,
+          height: 480,
+          hashFunction: 'keccak256(bytes)',
+          hash: '0xb115f2bf09994e79726db27a7b8d5a0de41a5b81d11b59b3038fa158718266ff',
+          url: 'ipfs://QmakaRZxJMMqwQFJY98J3wjbqYVDnaSZ9sEqBF9iMv3GNX',
+        },
+      ],
+      tags: ['public profile'],
+      links: [],
+    };
+
+    const encodedData = myERC725.encodeData({
+      LSP3Profile: {
+        json,
+        url: 'ifps://QmbKvCVEePiDKxuouyty9bMsWBAxZDGr2jhxd4pLGLx95D',
+      },
+    });
+
+    const decodedData = myERC725.decodeData({
+      LSP3Profile: encodedData.LSP3Profile.value,
+    });
+
+    assert.deepStrictEqual(
+      decodedData.LSP3Profile.url,
+      'ifps://QmbKvCVEePiDKxuouyty9bMsWBAxZDGr2jhxd4pLGLx95D',
+    );
+    assert.deepStrictEqual(
+      decodedData.LSP3Profile.hash,
+      hashData(json, SUPPORTED_HASH_FUNCTION_STRINGS.KECCAK256_UTF8),
+    );
+    assert.deepStrictEqual(
+      decodedData.LSP3Profile.hashFunction,
+      SUPPORTED_HASH_FUNCTION_STRINGS.KECCAK256_UTF8,
+    );
   });
 });
